requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt --no-annotate
agct==0.2.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
anyvar==1.0.0
asn1crypto==1.5.1
asttokens==3.0.1
attrs==25.4.0
biocommons-seqrepo==0.6.11
bioutils==0.6.1
boto3==1.42.29
botocore==1.42.29
canonicaljson==2.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
configparser==7.2.0
cryptography==46.0.3
decorator==5.2.1
dill==0.3.9
executing==2.2.1
fastapi==0.128.0
filelock==3.20.3
ga4gh-cat-vrs==0.7.2
ga4gh-va-spec==0.4.3
ga4gh-vrs==2.2.0
h11==0.16.0
hgvs==1.5.4
humanfriendly==10.0
idna==3.11
importlib-metadata==8.7.1
ipython==8.38.0
jedi==0.19.2
jmespath==1.0.1
matplotlib-inline==0.2.1
packaging==25.0
parsley==1.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.1
prompt-toolkit==3.0.52
psycopg2==2.9.11
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pyopenssl==25.3.0
pysam==0.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
snowflake-connector-python==4.2.0
snowflake-sqlalchemy==1.8.2
sortedcontainers==2.4.0
sqlalchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.50.0
tabulate==0.9.0
tomlkit==0.14.0
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
wags-tails==0.4.0
wcwidth==0.2.14
yoyo-migrations==9.0.0
zipp==3.23.0
